+++ playground/dynamic-import/nested/then-css/inner.js
@@ -1,0 +1,2 @@
+import './inner.css'
+export const inner = 'then-css-inner'