+++ playground/hmr/vite.config.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises'
 import path from 'node:path'
 import { defineConfig } from 'vite'
 import type { Plugin } from 'vite'
-import { TestCssLinkPlugin } from './css-link/plugin'
+import { TestCssLinkPlugin } from './css-link/plugin.ts'
 
 export default defineConfig(({ command }) => ({
   experimental: {