+++ playground/lib/vite.terser.config.js
@@ -1,6 +1,6 @@
 import path from 'node:path'
 import { defineConfig } from 'vite'
-import baseConfig from './vite.config'
+import baseConfig from './vite.config.js'
 
 export default defineConfig({
   ...baseConfig,